Devin Wolfe 11/27/2015 - 6:18 P.M.
Added the FIFO module instantiation to the top level module.  Added the
SysClk signal to the Tx FSM module for the control signals - now the FSM
itself is clocked on the fast system clock, but the transmit loop is
clocked on the slower baud clock.  This will increase responsiveness at
slower baud rates.

I also added a new interface for the UART.  It has two modports, one with
functions for the testbenchm and one with the full set of signals for the
UART module.  The read and write functions are currently only stubbed out.
Also, the too level UART module needs to have all of the interface
connections made before it will compile.  The module instantiations will
need to have their port assignments updated to point to the interface
ports.