Devin Wolfe
Fixed the BIST module - added an extra state to prevent the module from
causing a second, erroneous transmission after the test had completed, and
added the RTS and Tx_Busy signals to the port list.  If either RTS is not
asserted or Tx_Busy IS asserted, the BIST module will not start the test
process.